smart-form/requirements.txt
streamlit
openai
python-dotenv
sentence_transformers
numpy